src: use '\n' instead of std::endl unless flushing is desired
std::endl should not be used as a generic newline since as well as
outputting a '\n', it flushes the output stream.  This can severely
reduce i/o performance on large files.  For stdout, the output
is line-buffered by default.

NB: std::cout must be flushed before any call to system/fork