gen-cfg: add support for omitting entire tool config with a conditional
Conditionals at the start of a tool's option list are now evaluated
in the same way as options at the start of sublists.  If the conditional
evaluates to false, all subsequent options are skipped.

If the resulting configuration file would be empty (or only contains
comments), no file is output.  If there is an existing file, it is
deleted.