cfg/cfp: don't code using subdivision mesh when pp >= 0.24
The CfP-submitted bitstreams were not encoded using subdivision meshes
for rate points decimated to between 0.24 to 0.72.

Caveat: While the encoder doesn't use the subdivision mesh, the base
mesh it does use is determined using fitsubdiv with subdivision enabled
for decimation at 0.24 and 0.36.  i.e. the base mesh is fitted to the
(unused) subdivided mesh.